Move html parse to commonMain
