+++ backend/html-parse-data/src/commonMain/kotlin/com/gchristov/thecodinglove/htmlparsedata/HtmlParseDataModule.kt
@@ -2,10 +2,7 @@ package com.gchristov.thecodinglove.htmlparsedata
 
 import com.gchristov.thecodinglove.htmlparsedata.usecase.ParseHtmlPostsUseCase
 import com.gchristov.thecodinglove.htmlparsedata.usecase.ParseHtmlTotalPostsUseCase
-import com.gchristov.thecodinglove.htmlparsedata.usecase.RealParseHtmlPostsUseCase
-import com.gchristov.thecodinglove.htmlparsedata.usecase.RealParseHtmlTotalPostsUseCase
 import com.gchristov.thecodinglove.kmpcommonkotlin.di.DiModule
-import kotlinx.coroutines.Dispatchers
 import org.kodein.di.DI
 import org.kodein.di.bindProvider
 
@@ -18,12 +15,8 @@ object HtmlParseDataModule : DiModule() {
             bindProvider { provideParseHtmlPostsUseCase() }
         }
     }
-
-    private fun provideParseHtmlTotalPostsUseCase(): ParseHtmlTotalPostsUseCase = RealParseHtmlTotalPostsUseCase(
-        dispatcher = Dispatchers.Default
-    )
-
-    private fun provideParseHtmlPostsUseCase(): ParseHtmlPostsUseCase = RealParseHtmlPostsUseCase(
-        dispatcher = Dispatchers.Default
-    )
 }
+
+expect fun provideParseHtmlTotalPostsUseCase(): ParseHtmlTotalPostsUseCase
+
+expect fun provideParseHtmlPostsUseCase(): ParseHtmlPostsUseCase
