+++ backend/html-parse-data/src/commonMain/kotlin/com/gchristov/thecodinglove/htmlparsedata/usecase/ParseHtmlPostsUseCase.kt
@@ -1,0 +1,8 @@
+package com.gchristov.thecodinglove.htmlparsedata.usecase
+
+import arrow.core.Either
+import com.gchristov.thecodinglove.htmlparsedata.HtmlPost
+
+interface ParseHtmlPostsUseCase {
+    suspend operator fun invoke(html: String): Either<Throwable, List<HtmlPost>>
+}
