+++ backend/html-parse-data/src/commonMain/kotlin/com/gchristov/thecodinglove/htmlparsedata/usecase/ParseHtmlTotalPostsUseCase.kt
@@ -1,0 +1,7 @@
+package com.gchristov.thecodinglove.htmlparsedata.usecase
+
+import arrow.core.Either
+
+interface ParseHtmlTotalPostsUseCase {
+    suspend operator fun invoke(html: String): Either<Throwable, Int>
+}
